refactor: moved location impl. logic; replaced Location to EncounterLocation trait with regular function
